Pipeline/robo/requirements.txt
pyzmq
numpy 